Move views to views package
