+++ app/src/main/java/com/cianjansen/warofsuits/ui/views/PlayingCardView.kt
@@ -1,4 +1,4 @@
-package com.cianjansen.warofsuits.ui
+package com.cianjansen.warofsuits.ui.views
 
 import android.content.Context
 import android.util.AttributeSet
