+++ app/src/main/res/layout/activity_game.xml
@@ -6,7 +6,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     >
 
-    <com.cianjansen.warofsuits.ui.PlayingCardView
+    <com.cianjansen.warofsuits.ui.views.PlayingCardView
         android:id="@+id/pcv_yours"
         android:layout_width="@dimen/playing_card_width"
         android:layout_height="@dimen/playing_card_height"
@@ -26,7 +26,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         android:background="?colorPrimary"/>
 
-    <com.cianjansen.warofsuits.ui.PlayingCardView
+    <com.cianjansen.warofsuits.ui.views.PlayingCardView
         android:id="@+id/pcv_opponent"
         android:layout_width="@dimen/playing_card_width"
         android:layout_height="@dimen/playing_card_height"
